+++ packages/ts-transform/src/__tests__/import.ts
@@ -1,3 +1,0 @@
-import { content } from './imported'
-
-export default content