+++ packages/ts-transform/src/__tests__/require.ts
@@ -1,3 +1,0 @@
-const required = require('./required')
-
-export default required